processor/requirements.txt
datasets
datasets[vision]
python-dotenv